02/07/2025 (5)

+++ src/app/service/indexed-db.service.ts
@@ -1,4 +1,5 @@
-import { Injectable } from '@angular/core';
+import { isPlatformBrowser } from '@angular/common';
+import { Inject, Injectable, PLATFORM_ID } from '@angular/core';
 import { openDB } from 'idb';
 
 @Injectable({
@@ -6,10 +7,26 @@ import { openDB } from 'idb';
 })
 export class IndexedDbService {
   private dbName = 'AirlineDB';
+  isBrowser: any;
+  constructor(@Inject(PLATFORM_ID) platformId: Object) {
+    this.isBrowser = isPlatformBrowser(platformId);
+  }
 
-
+  async getData() {
+    if (!this.isBrowser) {
+      return Promise.resolve(null); // Alternative pour le serveur
+    }
+    
+    // Code IndexedDB uniquement côté client
+    return new Promise((resolve) => {
+      const request = indexedDB.open('MyDB');
+      request.onsuccess = () => resolve(request.result);
+    });
+  }
 
   async initDB() {
+    let response = await this.getData();
+    if(response == null) return Promise.resolve(null);; 
     return await openDB(this.dbName, 1, {
       upgrade(db) {
 
@@ -21,71 +38,86 @@ export class IndexedDbService {
   }
 
   async setUser(user: any) {
+    const db = await this.initDB();
+    if(db == null) return ; 
     try {
-      const db = await this.initDB();
+      
       await db.put('user', user);
     }catch {
-      const db = await this.initDB();
+      
       await db.put('user', user);
     }
   }
   async setRole(role :   string){
     const db = await this.initDB();
+    if(db == null) return ; 
+    
     await db.put('role',  role);
   }
 
   async  deleteRole(){
     const db = await this.initDB();
+    if(db == null) return ; 
+   
     await db.delete('role' , 0);
   }
 
   async getUser(  ) {
+    const db = await this.initDB();
+    if(db == null) return  null; 
     try{
-      const db = await this.initDB();
+       
       return (await db.getAll("user")).at(0);
     }catch {
-      const db = await this.initDB();
+    
       return (await db.getAll("user")).at(0);
     }
   }
   async getRole() {
+    const db = await this.initDB();
+    if(db == null) return ""; 
     try {
-      const db = await this.initDB();
+   
       return (await db.getAll("role")).at(0);
     }catch {
-      const db = await this.initDB();
+    
       return (await db.getAll("role")).at(0);
     }
   }
   async getAllUser( ) {
+    const db = await this.initDB();
+    if(db == null) return ; 
     try{
-      const db = await this.initDB();
+   
       return await db.getAll('user');
     }catch(e){
-      const db = await this.initDB();
+    
       return await db.getAll('user');
     }
   }
   async getAllRole( ) {
+    const db = await this.initDB();
+    if(db == null) return []; 
     try {
-      const db = await this.initDB();
+       
       return await db.getAll('role');
     }catch (e){
-      const db = await this.initDB();
+      
       return await db.getAll('role');
     }
   }
 
 
   async deleteUser(  id: string) {
+    const db = await this.initDB();
+    if(db == null) return ; 
    try{
-     const db = await this.initDB();
+     
      await db.delete('user' , 0);
    }catch (e){
-     const db = await this.initDB();
+    
      await db.delete('user' , 0);
    }
   }
-
 
 }
